requirements.txt
Flask
Flask-SQLAlchemy
pytest
gunicorn
psycopg2-binary
python-dotenv
flask_mail
transformers